Add more performance mods to 1.8.9
I added only these because I believe LF/Ornithe mods are currently out-of-scope. Nothririum seems to help quite a bit with performance, and so does EntityCulling (with the PolyPatcher one off). However, there appear to be sign rendering issues with it; the original suggests to toggle off "Smart Animations", and I most likely have them on. If so, I will disable it, and report back.

Apologies for such a convoluted pull request description.